dstring and sstring now use unsigned long (why was this not done before?)
git-svn-id: file:///source/coreland/corelib/trunk@691 6e7d7e42-f3d6-da11-80d8-00d0b73f4485

+++ dstring.h
@@ -6,24 +6,24 @@
 
 typedef struct dstring {
   char *s;
-  unsigned int len;
-  unsigned int a;
+  unsigned long len;
+  unsigned long a;
 } dstring;
 
-int dstring_init(dstring *, unsigned int);
+int dstring_init(dstring *, unsigned long);
 void dstring_free(dstring *);
 
 int dstring_cat(dstring *, const dstring *);
-int dstring_catb(dstring *, const char *, unsigned int);
+int dstring_catb(dstring *, const char *, unsigned long);
 int dstring_cats(dstring *, const char *);
 int dstring_cat0(dstring *);
-unsigned int dstring_0(dstring *);
+unsigned long dstring_0(dstring *);
 
-int dstring_cpyb(dstring *, const char *, unsigned int);
+int dstring_cpyb(dstring *, const char *, unsigned long);
 int dstring_cpys(dstring *, const char *);
 int dstring_copy(dstring *, const dstring *);
 
 void dstring_trunc(dstring *);
-void dstring_chop(dstring *, unsigned int);
+void dstring_chop(dstring *, unsigned long);
 
 #endif
