+++ dstring_0.c
@@ -1,6 +1,6 @@
 #include "dstring.h"
 
-unsigned int dstring_0(dstring *ds)
+unsigned long dstring_0(dstring *ds)
 {
   if (ds->len == ds->a) --ds->len;
   ds->s[ds->len] = 0;
