+++ sstring_cpys.c
@@ -1,6 +1,6 @@
 #include "sstring.h"
 
-unsigned int sstring_cpys(register sstring *ss, const char *s)
+unsigned long sstring_cpys(register sstring *ss, const char *s)
 {
   register const char* t;
   for (t = s;;) {
